pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thirteen-f"
version = "0.1.0"
description = "Deterministic 13F ingestion and analysis tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "httpx>=0.25",
    "lxml>=5.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
13f = "thirteen_f.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
